adding a --pkgvers option to specify the version of the resulting package
We sometimes need to re-build packages that are the same version of code
or build packages for two different OS X versions that have the same
version of the underlying code. Munki uses the version
in the package file to decide if it needs to update a package, so sometimes
we need to be able to bump a minor version to have the package re-install.

Signed-off-by: Marc Jordan <marc.jordan@disneyanimation.com>